+++ src/Common/FullName.php
@@ -18,10 +18,25 @@ class FullName
      * @readonly
      */
     public ?string $patronymic = null;
-    public function __construct(string  $name, ?string $surname = null, ?string $patronymic = null)
+
+    public function __construct(string $name, ?string $surname = null, ?string $patronymic = null)
     {
-        $this->name = $name;
+        $this->name = trim($name);
+
+        if ($surname !== null) {
+            $surname = trim($surname);
+        }
+        if ($surname === '') {
+            $surname = null;
+        }
         $this->surname = $surname;
+
+        if ($patronymic !== null) {
+            $patronymic = trim($patronymic);
+        }
+        if ($patronymic === '') {
+            $patronymic = null;
+        }
         $this->patronymic = $patronymic;
     }
 
@@ -35,5 +50,10 @@ class FullName
             'surname' => $this->surname,
             'patronymic' => $this->patronymic
         ];
+    }
+
+    public function equals(self $fullName): bool
+    {
+        return $this->name === $fullName->name && $this->surname === $fullName->surname && $this->patronymic === $fullName->patronymic;
     }
 }