+++ src/Common/Result/Contacts/ContactResult.php
@@ -1,0 +1,15 @@
+<?php
+
+declare(strict_types=1);
+
+namespace B24io\Loyalty\SDK\Common\Result\Contacts;
+
+use B24io\Loyalty\SDK\Core\Result\AbstractResult;
+
+class ContactResult extends AbstractResult
+{
+    public function getContact(): ContactItemResult
+    {
+        return new ContactItemResult($this->getCoreResponse()->getResponseData()->result);
+    }
+}