+++ src/Services/Admin/Contacts/Contacts.php
@@ -7,9 +7,8 @@ namespace B24io\Loyalty\SDK\Services\Admin\Contacts;
 use B24io\Loyalty\SDK\Common\FullName;
 use B24io\Loyalty\SDK\Common\Gender;
 use B24io\Loyalty\SDK\Common\Requests\ItemsOrder;
-use B24io\Loyalty\SDK\Common\Result\Cards\CardsResult;
 use B24io\Loyalty\SDK\Common\Result\Contacts\AddedContactResult;
-use B24io\Loyalty\SDK\Common\Result\Contacts\ContactItemResult;
+use B24io\Loyalty\SDK\Common\Result\Contacts\ContactResult;
 use B24io\Loyalty\SDK\Common\Result\Contacts\ContactsResult;
 use B24io\Loyalty\SDK\Core\Command;
 use B24io\Loyalty\SDK\Core\Credentials\Context;
@@ -29,20 +28,25 @@ class Contacts extends AbstractService
      * @param FullName $fullName
      * @param DateTimeZone $timezone
      * @param Gender $gender
-     * @param PhoneNumber $mobilePhone
+     * @param PhoneNumber|null $mobilePhone
      * @param DateTimeImmutable|null $birthdate
      * @param array<string, string> $externalIds
-     *@return AddedContactResult
+     * @return AddedContactResult
      */
     public function add(
         FullName           $fullName,
         DateTimeZone       $timezone,
         Gender             $gender,
-        PhoneNumber        $mobilePhone,
+        ?PhoneNumber       $mobilePhone = null,
         ?DateTimeImmutable $birthdate = null,
         array              $externalIds = []
     ): AddedContactResult
     {
+        $rawMobilePhone = null;
+        if ($mobilePhone !== null) {
+            $rawMobilePhone = $this->phoneNumberUtil->format($mobilePhone, PhoneNumberFormat::E164);
+        }
+
         return new AddedContactResult($this->core->call(
             new Command(
                 Context::admin(),
@@ -53,7 +57,7 @@ class Contacts extends AbstractService
                     'timezone' => $timezone->getName(),
                     'gender' => (string)$gender,
                     'birthday' => ($nullsafeBirthdate = $birthdate) ? $nullsafeBirthdate->format('Y.m.d') : null,
-                    'mobile_phone' => $this->phoneNumberUtil->format($mobilePhone, PhoneNumberFormat::E164),
+                    'mobile_phone' => $rawMobilePhone,
                     'external_ids' => $externalIds
                 ],
                 null,
@@ -62,16 +66,16 @@ class Contacts extends AbstractService
             )));
     }
 
-    public function getById(Uuid $id): ContactItemResult
+    public function getById(Uuid $id): ContactResult
     {
-        return new ContactItemResult(
+        return new ContactResult(
             $this->core->call(
                 new Command(
                     Context::admin(),
                     RequestMethodInterface::METHOD_GET,
                     sprintf('contacts/%s', $id->toRfc4122()),
                 )
-            )->getResponseData()->result
+            )
         );
     }