+++ src/Services/ServiceBuilderFactory.php
@@ -17,7 +17,12 @@ use Symfony\Contracts\HttpClient\HttpClientInterface;
 
 class ServiceBuilderFactory
 {
-    public static function createAdminRoleServiceBuilder(string               $apiEndpointUrl, string               $apiClientId, string               $apiAdminKey, LoggerInterface      $logger, ?HttpClientInterface $httpClient = null): AdminServiceBuilder
+    public static function createAdminRoleServiceBuilder(
+        string               $apiEndpointUrl,
+        string               $apiClientId,
+        string               $apiAdminKey,
+        LoggerInterface      $logger,
+        ?HttpClientInterface $httpClient = null): AdminServiceBuilder
     {
         return new AdminServiceBuilder(
             self::getCore(
@@ -76,7 +81,7 @@ class ServiceBuilderFactory
         return new CurlHttpClient(
             [
                 'http_version' => '2.0',
-                'timeout' => 60,
+                'timeout' => 30,
             ],
         );
     }